Pull the utility functions into utils.
Also modify the `compare_runs` to take an additional `linestyles`
argument which is a callable returning an iterator.
